layout/base/tests/test_image_layers.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that images that are the only item in ThebesLayers get put into ImageLayers</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="changeImage()">
<div>
<img id="image" src="./image_rgrg-256x256.png" style="-moz-transform: translatex(1px)"></img>
</div>
<pre id="test">
<script type="application/javascript">
if (!navigator.platform.startsWith("Linux")) {
  SimpleTest.expectAssertions(0, 3);
}

SimpleTest.waitForExplicitFinish();

var image = document.getElementById("image");
var lastPaintCount;
var utils = SpecialPowers.getDOMWindowUtils(window);

function changeImage() {
  utils.checkAndClearPaintedState(image);

  image.src = "./image_rrgg-256x256.png";
  image.onload = function() {
    lastPaintCount = window.mozPaintCount;
    checkDonePainting();
  }
}

// Wait until we've painted after the image loads, then force a paint and make
// sure the image doesn't get drawn.
function checkDonePainting() {
  if (window.mozPaintCount == lastPaintCount || utils.isMozAfterPaintPending) {
    setTimeout(checkDonePainting, 30);
    return;
  }

  lastPaintCount = window.mozPaintCount;
  document.body.style.background = "black";
  checkDone();
}

function checkDone() {
  if (window.mozPaintCount == lastPaintCount || utils.isMozAfterPaintPending) {
    setTimeout(checkDone, 30);
    return;
  }

  ok(!utils.checkAndClearPaintedState(image), "Should not paint any thebes layers for our image!");
  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
